+++ src/TestSuite/MockTrait.php
@@ -18,6 +18,7 @@ namespace MeTools\TestSuite;
 use Cake\Controller\ComponentRegistry;
 use Cake\Controller\Controller;
 use Cake\View\View;
+use PHPUnit\Framework\TestCase;
 
 /**
  * A mock trait
@@ -25,14 +26,50 @@ use Cake\View\View;
 trait MockTrait
 {
     /**
+     * Internal method to set off a test failure if a class does not exist
+     * @param string $className Class name
+     * @return void
+     * @throw \PHPUnit\Framework\AssertionFailedError
+     */
+    protected function _classExistsOrFail(string $className): void
+    {
+        class_exists($className) ?: $this->fail('Class `' . $className . '` does not exist');
+    }
+
+    /**
+     * Gets the alias name for which a test is being performed, starting from a
+     *  class or a `TestCase` class.
+     *
+     * Example: class `MyPlugin\Test\TestCase\Controller\PagesControllerTest`
+     *  will return the string `Pages`.
+     * @param string|object $class Class name as string or object
+     * @return string
+     * @since 2.19.9
+     * @throw \PHPUnit\Framework\AssertionFailedError If the class does not
+     *  exist or if it is impossible to get its alias
+     */
+    protected function getAlias($class): string
+    {
+        $class = is_object($class) ? get_class($class) : $class;
+        $this->_classExistsOrFail($class);
+
+        $alias = preg_replace('/^(\w+)(Cell|Controller|Table|Validator)(Test)?$/', '$1', get_class_short_name($class), -1, $count);
+        $count ?: $this->fail('Unable to get the alias for the `' . $class . '` class');
+
+        return $alias;
+    }
+
+    /**
      * Gets the alias for a controller
      * @param string $className Controller class name
      * @return string
-     * @throws \ReflectionException
+     * @deprecated Use instead `getAlias()`
      */
     protected function getControllerAlias(string $className): string
     {
-        return substr(get_class_short_name($className), 0, -10);
+        deprecationWarning('Deprecated. Use instead `getAlias()`');
+
+        return $this->getAlias($className);
     }
 
     /**
@@ -58,8 +95,7 @@ trait MockTrait
      */
     protected function getMockForController(string $className, ?array $methods = [], ?string $alias = null): object
     {
-        class_exists($className) ?: $this->fail('Class `' . $className . '` does not exist');
-        $alias = $alias ?: $this->getControllerAlias($className);
+        $alias = $alias ?: $this->getAlias($className);
 
         return $this->getMockBuilder($className)
             ->setConstructorArgs([null, null, $alias])
@@ -82,35 +118,71 @@ trait MockTrait
     }
 
     /**
-     * Gets the classname for which a test is being performed, starting from the
-     *  test class name.
+     * Gets the class name for which a test is being performed, starting from a
+     *  `TestCase` class.
      *
      * Example: class `MyPlugin\Test\TestCase\Controller\PagesControllerTest`
      *  will return the string `MyPlugin\Controller\PagesController`.
-     * @param object $testClass A test class as object or string
+     * @param \PHPUnit\Framework\TestCase $testClass A `TestCase` class
      * @return string|null The class name for which a test is being performed or
      *  `null` on failure
      * @since 2.18.0
      */
-    public function getOriginClassName(object $testClass): ?string
+    protected function getOriginClassName(TestCase $testClass): ?string
     {
-        $className = preg_replace('/^\\\\?(.+)Test\\\\TestCase\\\\(.+)Test$/', '\1\2', get_class($testClass), -1, $count);
+        $className = preg_replace('/^([\w\\\\]+)Test\\\\TestCase\\\\([\w\\\\]+)Test$/', '$1$2', get_class($testClass), -1, $count);
 
         return $count ? $className : null;
     }
 
     /**
-     * Gets the classname for which a test is being performed, starting from the
-     *  test class name. It fails if the class does not exist
-     * @param object $testClass A test class as object or string
+     * Gets the class name for which a test is being performed, starting from a
+     *  `TestCase` class.
+     *
+     * It fails if the class cannot be determined or it does not exist.
+     * @param \PHPUnit\Framework\TestCase $testClass A `TestCase` class
      * @return string The class name for which a test is being performed
      * @since 2.19.2
+     * @throw \PHPUnit\Framework\AssertionFailedError If the class does not
+     *  exist or if it is impossible to get its origin class name
      */
-    public function getOriginClassNameOrFail(object $testClass): string
+    protected function getOriginClassNameOrFail(TestCase $testClass): string
     {
         $className = $this->getOriginClassName($testClass);
-        $className ?: $this->fail(sprintf('Unable to get the classname for the `%s` class', get_class($testClass)));
+        $className ?: $this->fail('Unable to get the classname for the `' . get_class($testClass) . '` class');
+        $this->_classExistsOrFail($className);
 
         return $className;
+    }
+
+    /**
+     * Gets the table class name from an alias
+     * @param string $alias Alias name
+     * @param string $plugin Plugin name. If left blank, it will be self-determined
+     * @return string
+     * @since 2.19.9
+     */
+    protected function getTableClassNameFromAlias(string $alias, string $plugin = ''): string
+    {
+        $plugin = str_replace('/', '\\', $plugin ?: $this->getPluginName($this));
+
+        return $plugin . '\\Model\\Table\\' . $alias . 'Table';
+    }
+
+    /**
+     * Gets the classname for which a test is being performed, starting from a
+     *  `TestCase` class.
+     *
+     * Example: class `MyPlugin\MySubNamespace\Test\TestCase\MyExampleTest`
+     *  will return the string `MyPlugin/MySubNamespace`.
+     * @param \PHPUnit\Framework\TestCase $testClass A `TestCase` class
+     * @return string
+     * @since 2.19.9
+     */
+    protected function getPluginName(TestCase $testClass): string
+    {
+        $className = get_class($testClass);
+
+        return str_replace('\\', '/', substr($className, 0, strpos($className, '\\Test\\TestCase')));
     }
 }