+++ tests/TestCase/TestSuite/MockTraitTest.php
@@ -14,10 +14,14 @@ declare(strict_types=1);
  */
 namespace MeTools\Test\TestCase\TestSuite;
 
+use AnotherTestPlugin\MyPlugin\Test\TestCase\Controller\MyExampleControllerTest;
+use App\Controller\PagesController;
+use App\Model\Table\PostsTable;
+use App\Model\Validation\PostValidator;
+use App\View\Cell\MyExampleCell;
 use MeTools\TestSuite\TestCase;
 use PHPUnit\Framework\AssertionFailedError;
 use PHPUnit\Framework\MockObject\MockObject;
-use ReflectionException;
 
 /**
  * MockTraitTest class
@@ -25,11 +29,36 @@ use ReflectionException;
 class MockTraitTest extends TestCase
 {
     /**
+     * Tests for `getAlias()` method
+     * @test
+     */
+    public function testGetAlias()
+    {
+        $this->assertSame('Pages', $this->getAlias(new PagesController()));
+        $this->assertSame('Posts', $this->getAlias(PostsTable::class));
+        $this->assertSame('Post', $this->getAlias(PostValidator::class));
+        $this->assertSame('MyExample', $this->getAlias(MyExampleCell::class));
+
+        $this->assertSame('MyExample', $this->getAlias(new MyExampleControllerTest()));
+
+        //Class with no alias
+        $this->assertException(AssertionFailedError::class, function () {
+            $this->getAlias(\stdClass::class);
+        }, 'Unable to get the alias for the `stdClass` class');
+
+        //No existing class
+        $this->expectException(AssertionFailedError::class);
+        $this->expectExceptionMessage('Class `No\Existing\Class` does not exist');
+        $this->getAlias('No\Existing\Class');
+    }
+
+    /**
      * Tests for `getControllerAlias()` method
      * @test
      */
     public function testGetControllerAlias()
     {
+        $current = error_reporting(E_ALL & ~E_USER_DEPRECATED);
         foreach ([
             'App\Controller\PagesController',
             'App\Controller\Admin\PagesController',
@@ -38,9 +67,11 @@ class MockTraitTest extends TestCase
         ] as $class) {
             $this->assertEquals('Pages', $this->getControllerAlias($class));
         }
+        error_reporting($current);
 
-        $this->expectException(ReflectionException::class);
-        $this->getControllerAlias('App\NoExisting\Class');
+        $this->expectDeprecation();
+        $this->expectExceptionMessageMatches('/^Deprecated\. Use instead `getAlias\(\)`/');
+        $this->getControllerAlias('App\Controller\PagesController');
     }
 
     /**
@@ -91,7 +122,27 @@ class MockTraitTest extends TestCase
         $this->assertSame(TestCase::class, $this->getOriginClassNameOrFail(new TestCaseTest()));
 
         $this->expectException(AssertionFailedError::class);
-        $this->expectExceptionMessage('Unable to get the classname for the `stdClass` class');
-        $this->getOriginClassNameOrFail(new \stdClass());
+        $this->expectExceptionMessage('Class `AnotherTestPlugin\MyPlugin\Controller\MyExampleController` does not exist');
+        $this->getOriginClassNameOrFail(new MyExampleControllerTest());
+    }
+
+    /**
+     * Tests for `getPluginName()` method
+     * @test
+     */
+    public function testGetPluginName()
+    {
+        $this->assertSame('MeTools', $this->getPluginName(new TestCaseTest()));
+        $this->assertSame('AnotherTestPlugin/MyPlugin', $this->getPluginName(new MyExampleControllerTest()));
+    }
+
+    /**
+     * Tests for `getTableClassNameFromAlias()` method
+     * @test
+     */
+    public function testGetTableClassNameFromAlias()
+    {
+        $this->assertSame('MeTools\Model\Table\PostsTable', $this->getTableClassNameFromAlias('Posts'));
+        $this->assertSame('MyPlugin\SubNamespace\Model\Table\PostsTable', $this->getTableClassNameFromAlias('Posts', 'MyPlugin/SubNamespace'));
     }
 }