+++ tests/test_app/Plugin/AnotherTestPlugin/tests/TestCase/Controller/MyExampleControllerTest.php
@@ -1,0 +1,13 @@
+<?php
+declare(strict_types=1);
+
+namespace AnotherTestPlugin\MyPlugin\Test\TestCase\Controller;
+
+use PHPUnit\Framework\TestCase;
+
+/**
+ * An example class test
+ */
+class MyExampleControllerTest extends TestCase
+{
+}