+++ tests/test_app/TestApp/Application.php
@@ -1,17 +1,6 @@
 <?php
 declare(strict_types=1);
-/**
- * This file is part of cakephp-assets.
- *
- * Licensed under The MIT License
- * For full copyright and license information, please see the LICENSE.txt
- * Redistributions of files must retain the above copyright notice.
- *
- * @copyright   Copyright (c) Mirko Pagliai
- * @link        https://github.com/mirko-pagliai/cakephp-assets
- * @license     https://opensource.org/licenses/mit-license.php MIT License
- * @since       2.17.3
- */
+
 namespace App;
 
 use Cake\Http\BaseApplication;
@@ -20,10 +9,7 @@ use Cake\Routing\Middleware\RoutingMiddleware;
 use MeTools\Plugin as MeTools;
 
 /**
- * Application setup class.
- *
- * This defines the bootstrapping logic and middleware layers you
- * want to use in your application.
+ * Application setup class
  */
 class Application extends BaseApplication
 {