+++ tests/test_app/TestApp/View/Cell/MyExampleCell.php
@@ -1,0 +1,10 @@
+<?php
+declare(strict_types=1);
+
+namespace App\View\Cell;
+
+use Cake\View\Cell;
+
+class MyExampleCell extends Cell
+{
+}